Изменение компоновки Модификация отчета Вычисление рассчитываемых значений при старте приложения (во избежание ошибки с формированием отчета)

+++ TechnogenicSecurity/Models/Substance.cs
@@ -12,10 +12,10 @@ namespace TechnogenicSecurity.Models
         public double MolarMass { get; set; }
         public double LCLS { get; set; } //lower concentration limit of (flame) spread НКПР
         public double UCLS { get; set; } //upper concentration limit of (flame) spread ВКПР
-        public int GasExplosionEnergy { get; set; }         //энергия взрыва газа смеси
-        public int GasAirMixExplosionEnergy { get; set; }   //энергия взрыва газо-воздушной смеси
+        public double GasExplosionEnergy { get; set; }         //энергия взрыва газа смеси
+        public double GasAirMixExplosionEnergy { get; set; }   //энергия взрыва газо-воздушной смеси
         public double Density { get; set; } //плотность стехиометрической смеси
         public double HiddenVaporizationHeat { get; set; }
-        public int BoilingTemperature { get; set; }
+        public double BoilingTemperature { get; set; }
     }
 }
